[#18]feat: add authorization to the posts endpoints

+++ src/app/api/posts/[id]/route.ts
@@ -1,5 +1,8 @@
+import { ROLES } from '@/app/types/types';
+import { nextAuthSecret } from '@/app/utils/env';
 import { prisma } from '@/lib/prisma';
-import { NextResponse } from 'next/server';
+import { getToken } from 'next-auth/jwt';
+import { NextRequest, NextResponse } from 'next/server';
 
 /**
  * @swagger
@@ -18,25 +21,32 @@ import { NextResponse } from 'next/server';
  *         description: Success
  *       404:
  *         description: No post with the ID found
+ *       401:
+ *         description: Unauthorized
+ *       500:
+ *         description: Internal Server Error
  */
 export async function GET(
   request: Request,
   { params }: { params: { id: string } }
 ) {
-  const id = params.id;
-  const post = await prisma.post.findUnique({
-    where: {
-      id
-    }
-  });
-
-  if (!post) {
-    return new NextResponse('No post with the ID found', {
-      status: 404
+  try {
+    const id = params.id;
+    const post = await prisma.post.findUnique({
+      where: {
+        id
+      }
     });
-  }
 
-  return NextResponse.json(post);
+    if (!post) {
+      return new NextResponse('No post with the ID found', {
+        status: 404
+      });
+    }
+    return NextResponse.json(post);
+  } catch (error: any) {
+    return new NextResponse(error.message, { status: 500 });
+  }
 }
 
 /**
@@ -56,24 +66,38 @@ export async function GET(
  *         description: Success
  *       404:
  *         description: No post with the ID found
+ *       401:
+ *         description: Unauthorized
+ *       500:
+ *         description: Internal Server Error
  */
 export async function PATCH(
-  request: Request,
+  request: NextRequest,
   { params }: { params: { id: string } }
 ) {
-  const id = params.id;
-  let json = await request.json();
+  try {
+    const token = await getToken({ req: request, secret: nextAuthSecret });
 
-  const updated_post = await prisma.post.update({
-    where: { id },
-    data: json
-  });
+    if (token?.role === ROLES.ADMIN) {
+      const id = params.id;
+      let json = await request.json();
 
-  if (!updated_post) {
-    return new NextResponse('No post with the ID found', { status: 404 });
+      const updated_post = await prisma.post.update({
+        where: { id },
+        data: json
+      });
+
+      if (!updated_post) {
+        return new NextResponse('No post with the ID found', { status: 404 });
+      }
+
+      return NextResponse.json(updated_post);
+    }
+
+    return new NextResponse('Unauthorized', { status: 401 });
+  } catch (error: any) {
+    return new NextResponse(error.message, { status: 500 });
   }
-
-  return NextResponse.json(updated_post);
 }
 
 /**
@@ -91,22 +115,28 @@ export async function PATCH(
  *     responses:
  *       204:
  *         description: Success
+ *       401:
+ *         description: Unauthorized
  *       404:
  *         description: No post with ID found
  *       500:
  *        description: Internal Server Error
  */
 export async function DELETE(
-  request: Request,
+  request: NextRequest,
   { params }: { params: { id: string } }
 ) {
   try {
-    const id = params.id;
-    await prisma.post.delete({
-      where: { id }
-    });
+    const token = await getToken({ req: request, secret: nextAuthSecret });
+    if (token?.role === ROLES.ADMIN) {
+      const id = params.id;
+      await prisma.post.delete({
+        where: { id }
+      });
 
-    return new NextResponse(null, { status: 204 });
+      return new NextResponse(null, { status: 204 });
+    }
+    return new NextResponse('Unauthorized', { status: 401 });
   } catch (error: any) {
     if (error.code === 'P2025') {
       return new NextResponse('No post with the ID found', { status: 404 });
